requirements.txt
streamlit==1.32.2
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
matplotlib==3.8.3
